+++ src/commands/mod.rs
@@ -5,4 +5,5 @@ pub mod clean;
 pub mod cython;
 pub mod doctor;
 pub mod pull;
+pub mod python;
 pub mod run;